requirements.txt
aiohttp==3.7.4.post0
ansicolors==1.1.8
anyio==3.4.0
async-timeout==3.0.1
attrs==21.3.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
google-api-core==2.3.2
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.3
httplib2==0.20.2
httpx==0.21.1
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
multidict==5.2.0
mycroft-messagebus-client==0.9.5
packaging==21.3
paho-mqtt==1.6.1
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.1.0
pyparsing==3.0.6
pytest==7.0.1
python-espeak==0.6.3
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
tomli==2.0.1
twitchio==2.1.4
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.3
yarl==1.7.2
